+++ app/models/navigation_item.rb
@@ -1,6 +1,6 @@
 class NavigationItem < ActiveRecord::Base
   include Concerns::Sortable
-  sortable :navigation_id
+  sortable scope: :navigation_id
 
   include Concerns::Translatable
   translatable_fields :label, :path