+++ changelog.md
@@ -1,5 +1,6 @@
 0.0.12 - xxxx-xx-xx
 --
+* Bugfix: navigation item wasn't using the sortable scope.
 
 
 0.0.11 - 2015-12-05