+++ changelog.md
@@ -2,6 +2,8 @@
 --
 * Bugfix: navigation item wasn't using the sortable scope.
 * Bugfix: make sure content widgets can have camelcased model names.
+* Order redirects based on the times they've been used.
+* Add a catch all route with a basic controller that checks for 404's
 * The tagbox no longer triggers tagged item creation errors.